+++ packages/core/src/plugins/index.ts
@@ -1,1 +1,2 @@
+export * from './dedupe';
 export * from './sse';